+++ .eslintrc.js
@@ -1,0 +1,24 @@
+module.exports = {
+  env: {
+    node: true,
+    commonjs: true,
+    jest: true,
+    es6: true,
+  },
+  extends: "eslint:recommended",
+  overrides: [
+    {
+      env: {
+        node: true,
+      },
+      files: [".eslintrc.{js,cjs}"],
+      parserOptions: {
+        sourceType: "script",
+      },
+    },
+  ],
+  parserOptions: {
+    ecmaVersion: "latest",
+  },
+  rules: {},
+};
