+++ examples/HelloWorld/apps/Hello/widget/Hello.jsx
