+++ examples/HelloWorld/apps/Hello/widget/HelloModule.jsx
