+++ examples/HelloWorld/modules/hello.js
