+++ main.js
@@ -1,0 +1,3 @@
+const { run: run_cli } = require("./lib/cli.js");
+
+module.exports = { run_cli };
