+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "bos-workspace",
   "description": "BOS Workspace CLI",
-  "version": "0.0.1-alpha.0",
-  "main": "src/index.js",
+  "version": "0.0.1-alpha.1",
+  "main": "main.js",
   "author": "Achraf - sking.near",
   "license": "ISC",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/sekaiking/bos-workspace.git"
+    "url": "git+https://github.com/NEARBuilders/bos-workspace.git"
   },
   "keywords": [
     "near",
@@ -15,21 +15,26 @@
     "nearprotocol"
   ],
   "bin": {
-    "bos-workspace": "./src/index.js",
-    "bw": "./src/index.js"
+    "bos-workspace": "./src/bin/bos-workspace",
+    "bw": "./src/bin/bos-workspace"
   },
   "scripts": {
     "test": "jest",
-    "pretty": "prettier --write '**/*.{js}'"
+    "format": "prettier --write .",
+    "lint": "eslint ."
   },
   "dependencies": {
-    "bos-cli": "^0.3.2",
+    "bos-cli": "^0.3.6",
     "chokidar": "^3.5.3",
-    "commander": "^11.0.0",
+    "commander": "^11.1.0",
     "express": "^4.18.2",
-    "jest": "^29.6.2",
+    "glob": "^10.3.10",
+    "jest": "^29.7.0",
     "mock-fs": "^5.2.0",
     "near-cli-rs": "^0.4.3",
-    "prettier": "^3.0.1"
+    "prettier": "^3.1.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.53.0"
   }
 }
