--- templates/HelloWorld/apps/Hello/bos.config.json
@@ -1,6 +1,0 @@
-{
-    "appAccount": "dummy.near",
-    "aliases": {
-        "hello": "<h1>Hello, world! Have a nice day!</h1>"
-    }
-}
